pyproject.toml
[project]
name = "hello-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
]
